Get rid of pg application truncation warning
[#47979793]

The 0.15.0 version of the pg gem set the application name as the full
path to the application, `$0`, which was the full path to `rspec` when
running a subset of tests.  This caused warning NOTICEs to be printed
that the string would be truncated.  The truncation warning disappears
with version 0.15.1 of the pg gem because the full path is not used.